Refactoring FnApiRunner execution logic
