+++ sdks/python/apache_beam/runners/portability/fn_api_runner.py
@@ -36,7 +36,6 @@ import time
 import uuid
 from builtins import object
 from typing import TYPE_CHECKING
-from typing import Any
 from typing import Callable
 from typing import DefaultDict
 from typing import Dict
@@ -53,10 +52,8 @@ from typing import Union
 from typing import overload
 
 import grpc
-from typing_extensions import Protocol
 
 import apache_beam as beam  # pylint: disable=ungrouped-imports
-from apache_beam import coders
 from apache_beam.coders.coder_impl import create_InputStream
 from apache_beam.coders.coder_impl import create_OutputStream
 from apache_beam.metrics import metric
@@ -77,12 +74,12 @@ from apache_beam.portability.api import endpoints_pb2
 from apache_beam.runners import pipeline_context
 from apache_beam.runners import runner
 from apache_beam.runners.portability import artifact_service
+from apache_beam.runners.portability import fn_api_runner_execution
 from apache_beam.runners.portability import fn_api_runner_transforms
 from apache_beam.runners.portability import portable_metrics
+from apache_beam.runners.portability.fn_api_runner_execution import add_residuals_and_channel_splits_to_deferred_inputs
 from apache_beam.runners.portability.fn_api_runner_transforms import create_buffer_id
 from apache_beam.runners.portability.fn_api_runner_transforms import only_element
-from apache_beam.runners.portability.fn_api_runner_transforms import split_buffer_id
-from apache_beam.runners.portability.fn_api_runner_transforms import unique_name
 from apache_beam.runners.worker import bundle_processor
 from apache_beam.runners.worker import data_plane
 from apache_beam.runners.worker import sdk_worker
@@ -90,12 +87,8 @@ from apache_beam.runners.worker.channel_factory import GRPCChannelFactory
 from apache_beam.runners.worker.sdk_worker import _Future
 from apache_beam.runners.worker.statecache import StateCache
 from apache_beam.transforms import environments
-from apache_beam.transforms import trigger
-from apache_beam.transforms.window import GlobalWindow
-from apache_beam.transforms.window import GlobalWindows
 from apache_beam.utils import profiler
 from apache_beam.utils import proto_utils
-from apache_beam.utils import windowed_value
 from apache_beam.utils.thread_pool_executor import UnboundedThreadPoolExecutor
 
 if TYPE_CHECKING:
@@ -261,161 +254,6 @@ class BeamFnControlServicer(beam_fn_api_pb2_grpc.BeamFnControlServicer):
     _LOGGER.debug('Runner: Requests multiplexing info: %s',
                   [(str(req), worker) for req, worker
                    in self._req_worker_mapping.items()])
-
-
-class Buffer(Protocol):
-  def __iter__(self):
-    # type: () -> Iterator[bytes]
-    pass
-
-  def append(self, item):
-    # type: (bytes) -> None
-    pass
-
-
-class PartitionableBuffer(Buffer, Protocol):
-  def partition(self, n):
-    # type: (int) -> List[List[bytes]]
-    pass
-
-
-class _ListBuffer(List[bytes]):
-  """Used to support parititioning of a list."""
-  def partition(self, n):
-    # type: (int) -> List[List[bytes]]
-    return [self[k::n] for k in range(n)]
-
-
-class _GroupingBuffer(object):
-  """Used to accumulate groupded (shuffled) results."""
-  def __init__(self,
-               pre_grouped_coder,  # type: coders.Coder
-               post_grouped_coder,  # type: coders.Coder
-               windowing
-              ):
-    # type: (...) -> None
-    self._key_coder = pre_grouped_coder.key_coder()
-    self._pre_grouped_coder = pre_grouped_coder
-    self._post_grouped_coder = post_grouped_coder
-    self._table = collections.defaultdict(list)  # type: DefaultDict[bytes, List[Any]]
-    self._windowing = windowing
-    self._grouped_output = None  # type: Optional[List[List[bytes]]]
-
-  def append(self, elements_data):
-    # type: (bytes) -> None
-    if self._grouped_output:
-      raise RuntimeError('Grouping table append after read.')
-    input_stream = create_InputStream(elements_data)
-    coder_impl = self._pre_grouped_coder.get_impl()
-    key_coder_impl = self._key_coder.get_impl()
-    # TODO(robertwb): We could optimize this even more by using a
-    # window-dropping coder for the data plane.
-    is_trivial_windowing = self._windowing.is_default()
-    while input_stream.size() > 0:
-      windowed_key_value = coder_impl.decode_from_stream(input_stream, True)
-      key, value = windowed_key_value.value
-      self._table[key_coder_impl.encode(key)].append(
-          value if is_trivial_windowing
-          else windowed_key_value.with_value(value))
-
-  def partition(self, n):
-    # type: (int) -> List[List[bytes]]
-    """ It is used to partition _GroupingBuffer to N parts. Once it is
-    partitioned, it would not be re-partitioned with diff N. Re-partition
-    is not supported now.
-    """
-    if not self._grouped_output:
-      if self._windowing.is_default():
-        globally_window = GlobalWindows.windowed_value(
-            None,
-            timestamp=GlobalWindow().max_timestamp(),
-            pane_info=windowed_value.PaneInfo(
-                is_first=True,
-                is_last=True,
-                timing=windowed_value.PaneInfoTiming.ON_TIME,
-                index=0,
-                nonspeculative_index=0)).with_value
-        windowed_key_values = lambda key, values: [
-            globally_window((key, values))]
-      else:
-        # TODO(pabloem, BEAM-7514): Trigger driver needs access to the clock
-        #   note that this only comes through if windowing is default - but what
-        #   about having multiple firings on the global window.
-        #   May need to revise.
-        trigger_driver = trigger.create_trigger_driver(self._windowing, True)
-        windowed_key_values = trigger_driver.process_entire_key
-      coder_impl = self._post_grouped_coder.get_impl()
-      key_coder_impl = self._key_coder.get_impl()
-      self._grouped_output = [[] for _ in range(n)]
-      output_stream_list = []
-      for _ in range(n):
-        output_stream_list.append(create_OutputStream())
-      for idx, (encoded_key, windowed_values) in enumerate(self._table.items()):
-        key = key_coder_impl.decode(encoded_key)
-        for wkvs in windowed_key_values(key, windowed_values):
-          coder_impl.encode_to_stream(wkvs, output_stream_list[idx % n], True)
-      for ix, output_stream in enumerate(output_stream_list):
-        self._grouped_output[ix] = [output_stream.get()]
-      self._table.clear()
-    return self._grouped_output
-
-  def __iter__(self):
-    # type: () -> Iterator[bytes]
-    """ Since partition() returns a list of lists, add this __iter__ to return
-    a list to simplify code when we need to iterate through ALL elements of
-    _GroupingBuffer.
-    """
-    return itertools.chain(*self.partition(1))
-
-
-class _WindowGroupingBuffer(object):
-  """Used to partition windowed side inputs."""
-  def __init__(self,
-               access_pattern,
-               coder  # type: coders.WindowedValueCoder
-              ):
-    # type: (...) -> None
-    # Here's where we would use a different type of partitioning
-    # (e.g. also by key) for a different access pattern.
-    if access_pattern.urn == common_urns.side_inputs.ITERABLE.urn:
-      self._kv_extractor = lambda value: ('', value)
-      self._key_coder = coders.SingletonCoder('')  # type: coders.Coder
-      self._value_coder = coder.wrapped_value_coder
-    elif access_pattern.urn == common_urns.side_inputs.MULTIMAP.urn:
-      self._kv_extractor = lambda value: value
-      self._key_coder = coder.wrapped_value_coder.key_coder()
-      self._value_coder = (
-          coder.wrapped_value_coder.value_coder())
-    else:
-      raise ValueError(
-          "Unknown access pattern: '%s'" % access_pattern.urn)
-    self._windowed_value_coder = coder
-    self._window_coder = coder.window_coder
-    self._values_by_window = collections.defaultdict(list)  # type: DefaultDict[Tuple[str, BoundedWindow], List[Any]]
-
-  def append(self, elements_data):
-    # type: (bytes) -> None
-    input_stream = create_InputStream(elements_data)
-    while input_stream.size() > 0:
-      windowed_val_coder_impl = self._windowed_value_coder.get_impl()  # type: WindowedValueCoderImpl
-      windowed_value = windowed_val_coder_impl.decode_from_stream(
-          input_stream, True)
-      key, value = self._kv_extractor(windowed_value.value)
-      for window in windowed_value.windows:
-        self._values_by_window[key, window].append(value)
-
-  def encoded_items(self):
-    # type: () -> Iterator[Tuple[bytes, bytes, bytes]]
-    value_coder_impl = self._value_coder.get_impl()
-    key_coder_impl = self._key_coder.get_impl()
-    for (key, window), values in self._values_by_window.items():
-      encoded_window = self._window_coder.encode(window)
-      encoded_key = key_coder_impl.encode_nested(key)
-      output_stream = create_OutputStream()
-      for value in values:
-        value_coder_impl.encode_to_stream(value, output_stream, True)
-      yield encoded_key, encoded_window, output_stream.get()
-
 
 class FnApiRunner(runner.PipelineRunner):
 
@@ -590,7 +428,8 @@ class FnApiRunner(runner.PipelineRunner):
 
     try:
       with self.maybe_profile():
-        pcoll_buffers = collections.defaultdict(_ListBuffer)  # type: DefaultDict[bytes, PartitionableBuffer]
+        pcoll_buffers = collections.defaultdict(
+            fn_api_runner_execution._ListBuffer)  # type: DefaultDict[bytes, fn_api_runner_execution._ListBuffer]
         for stage in stages:
           stage_results = self._run_stage(
               worker_handler_manager.get_worker_handlers,
@@ -605,44 +444,6 @@ class FnApiRunner(runner.PipelineRunner):
       worker_handler_manager.close_all()
     return RunnerResult(
         runner.PipelineState.DONE, monitoring_infos_by_stage, metrics_by_stage)
-
-  def _store_side_inputs_in_state(self,
-                                  worker_handler,  # type: WorkerHandler
-                                  context,  # type: pipeline_context.PipelineContext
-                                  pipeline_components,  # type: beam_runner_api_pb2.Components
-                                  data_side_input,  # type: DataSideInput
-                                  pcoll_buffers,  # type: Mapping[bytes, PartitionableBuffer]
-                                  safe_coders
-                                 ):
-    # type: (...) -> None
-    for (transform_id, tag), (buffer_id, si) in data_side_input.items():
-      _, pcoll_id = split_buffer_id(buffer_id)
-      value_coder = context.coders[safe_coders[
-          pipeline_components.pcollections[pcoll_id].coder_id]]
-      elements_by_window = _WindowGroupingBuffer(si, value_coder)
-      for element_data in pcoll_buffers[buffer_id]:
-        elements_by_window.append(element_data)
-
-      if si.urn == common_urns.side_inputs.ITERABLE.urn:
-        for _, window, elements_data in elements_by_window.encoded_items():
-          state_key = beam_fn_api_pb2.StateKey(
-              iterable_side_input=beam_fn_api_pb2.StateKey.IterableSideInput(
-                  transform_id=transform_id,
-                  side_input_id=tag,
-                  window=window))
-          worker_handler.state.append_raw(state_key, elements_data)
-      elif si.urn == common_urns.side_inputs.MULTIMAP.urn:
-        for key, window, elements_data in elements_by_window.encoded_items():
-          state_key = beam_fn_api_pb2.StateKey(
-              multimap_side_input=beam_fn_api_pb2.StateKey.MultimapSideInput(
-                  transform_id=transform_id,
-                  side_input_id=tag,
-                  window=window,
-                  key=key))
-          worker_handler.state.append_raw(state_key, elements_data)
-      else:
-        raise ValueError(
-            "Unknown access pattern: '%s'" % si.urn)
 
   def _run_bundle_multiple_times_for_testing(
       self,
@@ -663,7 +464,7 @@ class FnApiRunner(runner.PipelineRunner):
       try:
         worker_handler.state.checkpoint()
         testing_bundle_manager = ParallelBundleManager(
-            worker_handler_list, lambda pcoll_id: _ListBuffer(),
+            worker_handler_list, lambda pcoll_id: fn_api_runner_execution._ListBuffer(),
             get_input_coder_callable, process_bundle_descriptor,
             self._progress_frequency, k,
             num_workers=self._num_workers,
@@ -707,59 +508,16 @@ class FnApiRunner(runner.PipelineRunner):
         for windowed_key_timer in timers_by_key_and_window.values():
           windowed_timer_coder_impl.encode_to_stream(
               windowed_key_timer, out, True)
-        deferred_inputs[transform_id] = _ListBuffer([out.get()])
-        written_timers[:] = []
-
-  def _add_residuals_and_channel_splits_to_deferred_inputs(
-      self,
-      splits,  # type: List[beam_fn_api_pb2.ProcessBundleSplitResponse]
-      get_input_coder_callable,
-      input_for_callable,
-      last_sent,
-      deferred_inputs  # type: DefaultDict[str, PartitionableBuffer]
-  ):
-    # type: (...) -> None
-
-    prev_stops = {}  # type: Dict[str, int]
-    for split in splits:
-      for delayed_application in split.residual_roots:
         deferred_inputs[
-            input_for_callable(
-                delayed_application.application.transform_id,
-                delayed_application.application.input_id)
-        ].append(delayed_application.application.element)
-      for channel_split in split.channel_splits:
-        coder_impl = get_input_coder_callable(channel_split.transform_id)
-        # TODO(SDF): This requires determanistic ordering of buffer iteration.
-        # TODO(SDF): The return split is in terms of indices.  Ideally,
-        # a runner could map these back to actual positions to effectively
-        # describe the two "halves" of the now-split range.  Even if we have
-        # to buffer each element we send (or at the very least a bit of
-        # metadata, like position, about each of them) this should be doable
-        # if they're already in memory and we are bounding the buffer size
-        # (e.g. to 10mb plus whatever is eagerly read from the SDK).  In the
-        # case of non-split-points, we can either immediately replay the
-        # "non-split-position" elements or record them as we do the other
-        # delayed applications.
-
-        # Decode and recode to split the encoded buffer by element index.
-        all_elements = list(coder_impl.decode_all(b''.join(last_sent[
-            channel_split.transform_id])))
-        residual_elements = all_elements[
-            channel_split.first_residual_element : prev_stops.get(
-                channel_split.transform_id, len(all_elements)) + 1]
-        if residual_elements:
-          deferred_inputs[channel_split.transform_id].append(
-              coder_impl.encode_all(residual_elements))
-        prev_stops[
-            channel_split.transform_id] = channel_split.last_primary_element
+            transform_id] = fn_api_runner_execution._ListBuffer([out.get()])
+        written_timers[:] = []
 
   @staticmethod
   def _extract_stage_data_endpoints(
       stage,  # type: fn_api_runner_transforms.Stage
       pipeline_components,  # type: beam_runner_api_pb2.Components
       data_api_service_descriptor,
-      pcoll_buffers  # type: DefaultDict[bytes, PartitionableBuffer]
+      pcoll_buffers  # type: DefaultDict[bytes, fn_api_runner_execution.PartitionableBuffer]
   ):
     # type: (...) -> Tuple[Dict[Tuple[str, str], PartitionableBuffer], DataSideInput, Dict[Tuple[str, str], bytes]]
 
@@ -775,7 +533,8 @@ class FnApiRunner(runner.PipelineRunner):
         if transform.spec.urn == bundle_processor.DATA_INPUT_URN:
           target = transform.unique_name, only_element(transform.outputs)
           if pcoll_id == fn_api_runner_transforms.IMPULSE_BUFFER:
-            data_input[target] = _ListBuffer([ENCODED_IMPULSE_VALUE])
+            data_input[target] = fn_api_runner_execution._ListBuffer(
+                [ENCODED_IMPULSE_VALUE])
           else:
             data_input[target] = pcoll_buffers[pcoll_id]
           coder_id = pipeline_components.pcollections[
@@ -804,7 +563,7 @@ class FnApiRunner(runner.PipelineRunner):
                  worker_handler_factory,  # type: Callable[[Optional[str], int], List[WorkerHandler]]
                  pipeline_components,  # type: beam_runner_api_pb2.Components
                  stage,  # type: fn_api_runner_transforms.Stage
-                 pcoll_buffers,  # type: DefaultDict[bytes, PartitionableBuffer]
+                 pcoll_buffers,  # type: DefaultDict[bytes, fn_api_runner_execution.PartitionableBuffer]
                  safe_coders
                 ):
     # type: (...) -> beam_fn_api_pb2.InstructionResponse
@@ -820,18 +579,6 @@ class FnApiRunner(runner.PipelineRunner):
         ``beam.PCollection``.
       safe_coders (dict): TODO
     """
-    def iterable_state_write(values, element_coder_impl):
-      # type: (...) -> bytes
-      token = unique_name(None, 'iter').encode('ascii')
-      out = create_OutputStream()
-      for element in values:
-        element_coder_impl.encode_to_stream(element, out, True)
-      worker_handler.state.append_raw(
-          beam_fn_api_pb2.StateKey(
-              runner=beam_fn_api_pb2.StateKey.Runner(key=token)),
-          out.get())
-      return token
-
     worker_handler_list = worker_handler_factory(
         stage.environment, self._num_workers)
 
@@ -839,7 +586,9 @@ class FnApiRunner(runner.PipelineRunner):
     # info from any worker_handler and read from the first worker_handler.
     worker_handler = next(iter(worker_handler_list))
     context = pipeline_context.PipelineContext(
-        pipeline_components, iterable_state_write=iterable_state_write)
+        pipeline_components,
+        iterable_state_write=fn_api_runner_execution.make_iterable_state_write(
+            worker_handler))
     data_api_service_descriptor = worker_handler.data_api_service_descriptor()
 
     _LOGGER.info('Running %s', stage.name)
@@ -863,107 +612,75 @@ class FnApiRunner(runner.PipelineRunner):
 
     # Store the required side inputs into state so it is accessible for the
     # worker when it runs this bundle.
-    self._store_side_inputs_in_state(worker_handler,
-                                     context,
-                                     pipeline_components,
-                                     data_side_input,
-                                     pcoll_buffers,
-                                     safe_coders)
-
-    def get_buffer(buffer_id):
-      # type: (bytes) -> PartitionableBuffer
-      """Returns the buffer for a given (operation_type, PCollection ID).
-
-      For grouping-typed operations, we produce a ``_GroupingBuffer``. For
-      others, we produce a ``_ListBuffer``.
-      """
-      kind, name = split_buffer_id(buffer_id)
-      if kind in ('materialize', 'timers'):
-        # If `buffer_id` is not a key in `pcoll_buffers`, it will be added by
-        # the `defaultdict`.
-        return pcoll_buffers[buffer_id]
-      elif kind == 'group':
-        # This is a grouping write, create a grouping buffer if needed.
-        if buffer_id not in pcoll_buffers:
-          original_gbk_transform = name
-          transform_proto = pipeline_components.transforms[
-              original_gbk_transform]
-          input_pcoll = only_element(list(transform_proto.inputs.values()))
-          output_pcoll = only_element(list(transform_proto.outputs.values()))
-          pre_gbk_coder = context.coders[safe_coders[
-              pipeline_components.pcollections[input_pcoll].coder_id]]
-          post_gbk_coder = context.coders[safe_coders[
-              pipeline_components.pcollections[output_pcoll].coder_id]]
-          windowing_strategy = context.windowing_strategies[
-              pipeline_components
-              .pcollections[output_pcoll].windowing_strategy_id]
-          pcoll_buffers[buffer_id] = _GroupingBuffer(
-              pre_gbk_coder, post_gbk_coder, windowing_strategy)
-      else:
-        # These should be the only two identifiers we produce for now,
-        # but special side input writes may go here.
-        raise NotImplementedError(buffer_id)
-      return pcoll_buffers[buffer_id]
-
-    def get_input_coder_impl(transform_id):
-      return context.coders[safe_coders[
-          beam_fn_api_pb2.RemoteGrpcPort.FromString(
-              process_bundle_descriptor.transforms[transform_id].spec.payload
-          ).coder_id
-      ]].get_impl()
+    fn_api_runner_execution.store_side_inputs_in_state(
+        worker_handler,
+        context,
+        pipeline_components,
+        data_side_input,
+        pcoll_buffers,
+        safe_coders)
 
     # Change cache token across bundle repeats
     cache_token_generator = FnApiRunner.get_cache_token_generator(static=False)
 
+    coder_getter_fn = fn_api_runner_execution.make_input_coder_getter(
+        context, process_bundle_descriptor, safe_coders)
+
     self._run_bundle_multiple_times_for_testing(
         worker_handler_list, process_bundle_descriptor, data_input, data_output,
-        get_input_coder_impl, cache_token_generator=cache_token_generator)
+        coder_getter_fn,
+        cache_token_generator=cache_token_generator)
 
     bundle_manager = ParallelBundleManager(
-        worker_handler_list, get_buffer, get_input_coder_impl,
+        worker_handler_list,
+        fn_api_runner_execution.make_input_buffer_fetcher(context,
+                                                          pcoll_buffers,
+                                                          pipeline_components,
+                                                          safe_coders),
+        coder_getter_fn,
         process_bundle_descriptor, self._progress_frequency,
         num_workers=self._num_workers,
         cache_token_generator=cache_token_generator)
 
     result, splits = bundle_manager.process_bundle(data_input, data_output)
 
-    def input_for(transform_id, input_id):
-      # type: (str, str) -> str
-      input_pcoll = process_bundle_descriptor.transforms[
-          transform_id].inputs[input_id]
-      for read_id, proto in process_bundle_descriptor.transforms.items():
-        if (proto.spec.urn == bundle_processor.DATA_INPUT_URN
-            and input_pcoll in proto.outputs.values()):
-          return read_id
-      raise RuntimeError(
-          'No IO transform feeds %s' % transform_id)
-
     last_result = result
     last_sent = data_input
 
     while True:
-      deferred_inputs = collections.defaultdict(_ListBuffer)  # type: DefaultDict[str, PartitionableBuffer]
+      deferred_inputs = collections.defaultdict(
+          fn_api_runner_execution._ListBuffer)  # type: DefaultDict[str, fn_api_runner_execution.PartitionableBuffer]
 
       self._collect_written_timers_and_add_to_deferred_inputs(
-          context, pipeline_components, stage, get_buffer, deferred_inputs)
+          context, pipeline_components, stage,
+          fn_api_runner_execution.make_input_buffer_fetcher(
+              context, pcoll_buffers,
+              pipeline_components, safe_coders),
+          deferred_inputs)
 
       # Queue any process-initiated delayed bundle applications.
       for delayed_application in last_result.process_bundle.residual_roots:
         deferred_inputs[
-            input_for(
+            fn_api_runner_execution.get_input_operation_name(
+                process_bundle_descriptor,
                 delayed_application.application.transform_id,
                 delayed_application.application.input_id)
         ].append(delayed_application.application.element)
 
       # Queue any runner-initiated delayed bundle applications.
-      self._add_residuals_and_channel_splits_to_deferred_inputs(
-          splits, get_input_coder_impl, input_for, last_sent, deferred_inputs)
+      add_residuals_and_channel_splits_to_deferred_inputs(
+          process_bundle_descriptor,
+          splits,
+          coder_getter_fn,
+          last_sent,
+          deferred_inputs)
 
       if deferred_inputs:
         # The worker will be waiting on these inputs as well.
         for other_input in data_input:
           if other_input not in deferred_inputs:
-            deferred_inputs[other_input] = _ListBuffer([])
+            deferred_inputs[
+                other_input] = fn_api_runner_execution._ListBuffer([])
         # TODO(robertwb): merge results
         # We cannot split deferred_input until we include residual_roots to
         # merged results. Without residual_roots, pipeline stops earlier and we
@@ -990,9 +707,9 @@ class FnApiRunner(runner.PipelineRunner):
   def _extract_endpoints(stage,  # type: fn_api_runner_transforms.Stage
                          pipeline_components,  # type: beam_runner_api_pb2.Components
                          data_api_service_descriptor, # type: Optional[endpoints_pb2.ApiServiceDescriptor]
-                         pcoll_buffers  # type: DefaultDict[bytes, PartitionableBuffer]
+                         pcoll_buffers  # type: DefaultDict[bytes, fn_api_runner_execution.PartitionableBuffer]
                         ):
-    # type: (...) -> Tuple[Dict[str, PartitionableBuffer], DataSideInput, DataOutput]
+    # type: (...) -> Tuple[Dict[str, fn_api_runner_execution.PartitionableBuffer], DataSideInput, DataOutput]
     """Returns maps of transform names to PCollection identifiers.
 
     Also mutates IO stages to point to the data ApiServiceDescriptor.
@@ -1011,7 +728,7 @@ class FnApiRunner(runner.PipelineRunner):
         PCollection buffer; `data_output` is a dictionary mapping
         (transform_name, output_name) to a PCollection ID.
     """
-    data_input = {}  # type: Dict[str, PartitionableBuffer]
+    data_input = {}  # type: Dict[str, fn_api_runner_execution.PartitionableBuffer]
     data_side_input = {}  # type: DataSideInput
     data_output = {}  # type: DataOutput
     for transform in stage.transforms:
@@ -1020,8 +737,9 @@ class FnApiRunner(runner.PipelineRunner):
         pcoll_id = transform.spec.payload
         if transform.spec.urn == bundle_processor.DATA_INPUT_URN:
           if pcoll_id == fn_api_runner_transforms.IMPULSE_BUFFER:
-            data_input[transform.unique_name] = _ListBuffer(
-                [ENCODED_IMPULSE_VALUE])
+            data_input[
+                transform.unique_name] = fn_api_runner_execution._ListBuffer(
+                    [ENCODED_IMPULSE_VALUE])
           else:
             data_input[transform.unique_name] = pcoll_buffers[pcoll_id]
           coder_id = pipeline_components.pcollections[
@@ -1098,7 +816,7 @@ class FnApiRunner(runner.PipelineRunner):
         self._overlay[self._key].append(item)
 
     StateType = Union[CopyOnWriteState,
-                      DefaultDict[bytes, Buffer]]
+                      DefaultDict[bytes, fn_api_runner_execution.Buffer]]
 
     def __init__(self):
       # type: () -> None
@@ -1976,7 +1694,7 @@ class BundleManager(object):
 
   def _generate_splits_for_testing(self,
                                    split_manager,
-                                   inputs,  # type: Mapping[str, PartitionableBuffer]
+                                   inputs,  # type: Mapping[str, fn_api_runner_execution.PartitionableBuffer]
                                    process_bundle_id):
     # type: (...) -> List[beam_fn_api_pb2.ProcessBundleSplitResponse]
     split_results = []  # type: List[beam_fn_api_pb2.ProcessBundleSplitResponse]
@@ -2038,7 +1756,7 @@ class BundleManager(object):
     return split_results
 
   def process_bundle(self,
-                     inputs,  # type: Mapping[str, PartitionableBuffer]
+                     inputs,  # type: Mapping[str, fn_api_runner_execution.PartitionableBuffer]
                      expected_outputs  # type: DataOutput
                     ):
     # type: (...) -> BundleProcessResult
@@ -2126,7 +1844,7 @@ class ParallelBundleManager(BundleManager):
     self._num_workers = kwargs.pop('num_workers', 1)
 
   def process_bundle(self,
-                     inputs,  # type: Mapping[str, PartitionableBuffer]
+                     inputs,  # type: Mapping[str, fn_api_runner_execution.PartitionableBuffer]
                      expected_outputs  # type: DataOutput
                     ):
     # type: (...) -> BundleProcessResult
